requirements.txt
requests==2.31.0
numpy>=1.25.0
pandas
fastapi>=0.95.0
uvicorn>=0.21.0
jinja2>=3.1.2
sqlalchemy>=2.0.0
pymysql>=1.0.3
python-jose[cryptography]
python-dotenv>=1.0.0
pydantic>=1.10.0
pydantic-settings
python-multipart
pytest>=7.0
alembic
itsdangerous

# Starlette and related dependencies
starlette>=0.27.0
httpx>=0.24.0

# Visual testing dependencies
selenium>=4.15.0
pillow>=10.0.0

# Additional testing dependencies
pytest-asyncio>=0.21.0
